Make internal config parsing helper functions private
