+++ src/neo4j/_async/io/_pool.py
@@ -31,6 +31,7 @@ from dataclasses import dataclass
 from logging import getLogger
 from random import choice
 
+from ..._api import check_access_mode
 from ..._async_compat.concurrency import (
     AsyncCondition,
     AsyncCooperativeRLock,
@@ -45,10 +46,7 @@ from ..._deadline import (
 )
 from ..._exceptions import BoltError
 from ..._routing import RoutingTable
-from ...api import (
-    check_access_mode,
-    READ_ACCESS,
-)
+from ...api import READ_ACCESS
 from ...exceptions import (
     ConfigurationError,
     ConnectionAcquisitionTimeoutError,
