+++ src/neo4j/_sync/driver.py
@@ -34,6 +34,8 @@ from .._api import (
     DRIVER_BOLT,
     DRIVER_NEO4J,
     NotificationMinimumSeverity,
+    parse_neo4j_uri,
+    parse_routing_context,
     RoutingControl,
     SECURITY_TYPE_SECURE,
     SECURITY_TYPE_SELF_SIGNED_CERTIFICATE,
@@ -63,8 +65,6 @@ from ..api import (
     Auth,
     BookmarkManager,
     Bookmarks,
-    parse_neo4j_uri,
-    parse_routing_context,
     READ_ACCESS,
     ServerInfo,
     TRUST_ALL_CERTIFICATES,
