+++ src/neo4j/api.py
@@ -20,13 +20,6 @@ from __future__ import annotations
 
 import abc
 import typing as t
-from urllib.parse import (
-    parse_qs,
-    urlparse,
-)
-
-from . import _api
-from .exceptions import ConfigurationError
 
 
 if t.TYPE_CHECKING:
@@ -65,11 +58,8 @@ __all__ = [
     "ServerInfo",
     "basic_auth",
     "bearer_auth",
-    "check_access_mode",
     "custom_auth",
     "kerberos_auth",
-    "parse_neo4j_uri",
-    "parse_routing_context",
 ]
 
 
@@ -443,96 +433,3 @@ class AsyncBookmarkManager(_Protocol, metaclass=abc.ABCMeta):
     async def get_bookmarks(self) -> t.Collection[str]: ...
 
     get_bookmarks.__doc__ = BookmarkManager.get_bookmarks.__doc__
-
-
-# TODO: 6.0 - make this function private
-def parse_neo4j_uri(uri):
-    parsed = urlparse(uri)
-
-    if parsed.username:
-        raise ConfigurationError("Username is not supported in the URI")
-
-    if parsed.password:
-        raise ConfigurationError("Password is not supported in the URI")
-
-    if parsed.scheme == URI_SCHEME_BOLT_ROUTING:
-        raise ConfigurationError(
-            f"Uri scheme {parsed.scheme!r} has been renamed. "
-            f"Use {URI_SCHEME_NEO4J!r}"
-        )
-    elif parsed.scheme == URI_SCHEME_BOLT:
-        driver_type = _api.DRIVER_BOLT
-        security_type = _api.SECURITY_TYPE_NOT_SECURE
-    elif parsed.scheme == URI_SCHEME_BOLT_SELF_SIGNED_CERTIFICATE:
-        driver_type = _api.DRIVER_BOLT
-        security_type = _api.SECURITY_TYPE_SELF_SIGNED_CERTIFICATE
-    elif parsed.scheme == URI_SCHEME_BOLT_SECURE:
-        driver_type = _api.DRIVER_BOLT
-        security_type = _api.SECURITY_TYPE_SECURE
-    elif parsed.scheme == URI_SCHEME_NEO4J:
-        driver_type = _api.DRIVER_NEO4J
-        security_type = _api.SECURITY_TYPE_NOT_SECURE
-    elif parsed.scheme == URI_SCHEME_NEO4J_SELF_SIGNED_CERTIFICATE:
-        driver_type = _api.DRIVER_NEO4J
-        security_type = _api.SECURITY_TYPE_SELF_SIGNED_CERTIFICATE
-    elif parsed.scheme == URI_SCHEME_NEO4J_SECURE:
-        driver_type = _api.DRIVER_NEO4J
-        security_type = _api.SECURITY_TYPE_SECURE
-    else:
-        supported_schemes = [
-            URI_SCHEME_BOLT,
-            URI_SCHEME_BOLT_SELF_SIGNED_CERTIFICATE,
-            URI_SCHEME_BOLT_SECURE,
-            URI_SCHEME_NEO4J,
-            URI_SCHEME_NEO4J_SELF_SIGNED_CERTIFICATE,
-            URI_SCHEME_NEO4J_SECURE,
-        ]
-        raise ConfigurationError(
-            f"URI scheme {parsed.scheme!r} is not supported. "
-            f"Supported URI schemes are {supported_schemes}. "
-            "Examples: bolt://host[:port] or "
-            "neo4j://host[:port][?routing_context]"
-        )
-
-    return driver_type, security_type, parsed
-
-
-# TODO: 6.0 - make this function private
-def check_access_mode(access_mode):
-    if access_mode not in {READ_ACCESS, WRITE_ACCESS}:
-        raise ValueError(
-            f"Unsupported access mode {access_mode}, must be one of "
-            f"'{READ_ACCESS}' or '{WRITE_ACCESS}'."
-        )
-
-    return access_mode
-
-
-# TODO: 6.0 - make this function private
-def parse_routing_context(query):
-    """
-    Parse the query portion of a URI.
-
-    Generates a routing context dictionary.
-    """
-    if not query:
-        return {}
-
-    context = {}
-    parameters = parse_qs(query, True)
-    for key in parameters:
-        value_list = parameters[key]
-        if len(value_list) != 1:
-            raise ConfigurationError(
-                f"Duplicated query parameters with key '{key}', value "
-                f"'{value_list}' found in query string '{query}'"
-            )
-        value = value_list[0]
-        if not value:
-            raise ConfigurationError(
-                f"Invalid parameters:'{key}={value}' in query string "
-                f"'{query}'."
-            )
-        context[key] = value
-
-    return context
